Add Tapto Card Reader v1.1
This PR adds a new Tapto Card Reader based off of RETROCASTLE's original one.

This Card Reader features a USB-C connector and has yellow LEDs on it.

There are two versions of the Card Reader:
- With 4pin right angle connector
- Without 4pin right angle connector

This includes EasyEDA source files as well as Gerber, BOM and PCL files for both versions of the board.